renamed gpt_thread.py to chat_thread.py
